requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.38.2
botocore==1.38.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
Django==4.2.20
django-rest-authtoken==2.1.4
django-storages==1.14.6
djangorestframework==3.16.0
docstring_parser==0.16
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
google-api-core==2.25.0rc0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kfp==2.12.2
kfp-pipeline-spec==0.6.0
kfp-server-api==2.4.0
kombu==5.5.3
kubernetes==30.1.0
oauthlib==3.2.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
